+++ src/roq/samples/python/flags/flags.cpp
@@ -8,50 +8,28 @@
 
 ABSL_FLAG(  //
     std::string,
-    deribit_exchange,
+    exchange,
     "deribit",
-    "name of the deribit exchange");
+    "exchange");
 
 ABSL_FLAG(  //
     std::string,
-    deribit_symbols,
-    "BTC-\\d{2}\\D{3}\\d{2}",  // e.g. "BTC-27MAR20"
-    "regex used to subscribe deribit symbols");
-
-ABSL_FLAG(  //
-    std::string,
-    coinbase_pro_exchange,
-    "coinbase-pro",
-    "name of the coinbase-pro exchange");
-
-ABSL_FLAG(  //
-    std::string,
-    coinbase_pro_symbols,
-    "BTC-.*",  // e.g. "BTC-USD"
-    "regex used to subscribe coinbase-pro symbols");
+    symbols,
+    "BTC-PERPETUAL",
+    "regex used to subscribe symbols");
 
 namespace roq {
 namespace samples {
 namespace python {
 namespace flags {
 
-std::string_view Flags::deribit_exchange() {
-  static const std::string result = absl::GetFlag(FLAGS_deribit_exchange);
+std::string_view Flags::exchange() {
+  static const std::string result = absl::GetFlag(FLAGS_exchange);
   return result;
 }
 
-std::string_view Flags::deribit_symbols() {
-  static const std::string result = absl::GetFlag(FLAGS_deribit_symbols);
-  return result;
-}
-
-std::string_view Flags::coinbase_pro_exchange() {
-  static const std::string result = absl::GetFlag(FLAGS_coinbase_pro_exchange);
-  return result;
-}
-
-std::string_view Flags::coinbase_pro_symbols() {
-  static const std::string result = absl::GetFlag(FLAGS_coinbase_pro_symbols);
+std::string_view Flags::symbols() {
+  static const std::string result = absl::GetFlag(FLAGS_symbols);
   return result;
 }