+++ src/roq/samples/python/foo.py
@@ -1,0 +1,3 @@
+def bar(bp, ap):
+    print(f"{bp},{ap}")
+    return 0.5 * (bp + ap)